+++ backend/danswer/auth/schemas.py
@@ -1,6 +1,5 @@
 import uuid
 from enum import Enum
-from typing import Optional
 
 from fastapi_users import schemas
 
@@ -34,9 +33,9 @@ class UserRead(schemas.BaseUser[uuid.UUID]):
 
 class UserCreate(schemas.BaseUserCreate):
     role: UserRole = UserRole.BASIC
-    has_web_login: Optional[bool] = True
+    has_web_login: bool | None = True
 
 
 class UserUpdate(schemas.BaseUserUpdate):
     role: UserRole
-    has_web_login: Optional[bool] = True
+    has_web_login: bool | None = True
