+++ backend/danswer/db/users.py
@@ -1,9 +1,11 @@
 from collections.abc import Sequence
 from uuid import UUID
 
+from fastapi_users.password import PasswordHelper
 from sqlalchemy import select
 from sqlalchemy.orm import Session
 
+from danswer.auth.schemas import UserRole
 from danswer.db.models import User
 
 
@@ -29,4 +31,23 @@ def get_user_by_email(email: str, db_session: Session) -> User | None:
 def fetch_user_by_id(db_session: Session, user_id: UUID) -> User | None:
     user = db_session.query(User).filter(User.id == user_id).first()  # type: ignore
 
+    return user
+
+
+def add_user_if_not_exists(email: str, db_session: Session) -> User:
+    user = get_user_by_email(email, db_session)
+    if user is not None:
+        return user
+
+    fastapi_users_pw_helper = PasswordHelper()
+    password = fastapi_users_pw_helper.generate()
+    hashed_pass = fastapi_users_pw_helper.hash(password)
+    user = User(
+        email=email,
+        hashed_password=hashed_pass,
+        has_web_login=False,
+        role=UserRole.BASIC,
+    )
+    db_session.add(user)
+    db_session.commit()
     return user
